Gadget System
At the moment, the system allows to change the values of: life, speed, damage (melee and ranged). The values should be within the range of 0 - 1 as they represent percentages. It is still necessary to create a system for them to be stored on the player, especially when changing levels

+++ Assets/_Scripts/UI/HealthManager.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class HealthManager : MonoBehaviour
 {
     private Player player;
-    private int playerLife;
+    private float playerLife;
     public Slider slider;
 
     // Start is called before the first frame update
